broker/pyproject.toml
[tool.poetry]
name = "broker"
version = "0.1.0"
description = "Message Broker Service"
authors = [
    "joefox@gmail.com"
]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13"
pika = "^1.3.2"
pydantic = "^2.11.1"
protobuf = "^5.26.1"
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
